moved open file in editor logic to BaseRemoteCommand
