+++ VSRAD.Package/Commands/DisassemblyCommand.cs
@@ -60,31 +60,5 @@ namespace VSRAD.Package.Commands
                 await ClearStatusBarAsync();
             }
         }
-
-        private void OpenFileInEditor(string path, string lineMarker)
-        {
-            OpenFileInEditor(path);
-            if (string.IsNullOrEmpty(lineMarker)) return;
-
-            var lineNumber = GetMarkedLineNumber(path, lineMarker);
-
-            var textManager = _serviceProvider.GetService(typeof(SVsTextManager)) as IVsTextManager2;
-            Assumes.Present(textManager);
-
-            textManager.GetActiveView2(1, null, (uint)_VIEWFRAMETYPE.vftCodeWindow, out var activeView);
-            activeView.SetCaretPos(lineNumber, 0);
-        }
-
-        private int GetMarkedLineNumber(string file, string lineMarker)
-        {
-            var lineNumber = 0;
-            foreach (var line in File.ReadLines(file))
-            {
-                if (line == lineMarker)
-                    return lineNumber;
-                ++lineNumber;
-            }
-            return 0;
-        }
     }
 }
